create admin panel

+++ apps/movies/models.py
@@ -56,10 +56,10 @@ class Movie(models.Model):
     directors = models.ManyToManyField(Actor, verbose_name="актёры", related_name="film_director")
     actors = models.ManyToManyField(Genre, verbose_name="жанры")
     world_premiere = models.DateField("Премьера в мире", default=date.today)
-    budget = models.PositiveSmallIntegerField("Бюджет", default=0, help_text="указывать сумму в долларах США")
-    fees_in_usa = models.PositiveSmallIntegerField(
+    budget = models.PositiveIntegerField("Бюджет", default=0, help_text="указывать сумму в долларах США")
+    fees_in_usa = models.PositiveIntegerField(
         "Сборы в США", default=0, help_text="указывать сумму в долларах США")
-    fees_in_world = models.PositiveSmallIntegerField(
+    fees_in_world = models.PositiveIntegerField(
         "Сборы в мире", default=0, help_text="указывать сумму в долларах США")
     category = models.ForeignKey(Category, verbose_name="Категория", on_delete=models.SET_NULL, null=True)
     url = models.SlugField(max_length=160, unique=True)
